Hide some links from unauthenticated users
Random visitors will no longer be tempted to try to delete, edit or
create items and categories. Once user permissions are added, they'll
need to be plugged into this as a second layer.

Moving these into partials is also a very good idea, and would make my
life easier if I could figure out a simple way to do it.