Adding more features ...
This is midway through the process. A lot of things are partially
implemented. Overall it's functional, but definitely not done.

* Menus are stored in-database now! But there's no interface to
configure whether categories appear in them, or to edit the headings.
* / and /about are partially in database, but the displayed text isn't
drawn from the database.
* Seeds are now implemented to populate the initial state of the db,
including an administrative user.
* The basics of a registration token system (you need a token to
register, the admin can generate tokens, and tokens can only be used
once) exists, but isn't attached to anything.