Improved menus, added flash
Menus are now stored in a much more logical way: in their own table!
That's just for them. Need to clean up the Core table somewhat; I'm
considering how it would be best to structure it.

Also added a bit of code so that flash notifications are actually
displayed; I'll work on making it a bit prettier, and adding more flash
notifications, later.