Merged all CSS
Since AJAX doesn't reload the entirety of a page, styles that weren't
part of the first page loaded were going missing. I've merged all of the
CSS together (or at least, it will be rendered as a single file) to
prevent any problems.

Oh, and also deleted some files that weren't being used.