Added tagging to items
Items can now be tagged, in addition to being categorized. Not a huge
feature, but a fun one. Depends on two things: the ruby gem
ActsAsTaggableOn and the jQuery Tags Input Plugin. Since the latter
doesn't have an included license, and so it isn't included in the files
in this repository.