Notifications with Toastr&Mailboxer
Set up notifications for most common events, using a mixture of
Mailboxer (for backend) and Toastr (to display stuff). Also set it up so
that loaded pages check for new notifications every 10 seconds; this is
potentially useless, but could come in handy depending on which
direction I take things in.