Reverted to using flash and toastr for notifications
Using Mailboxer for those notifications was a bit wasteful and
pointless; the flash is perfect for them, so using the flash is what
I'll do. Still going to keep Mailboxer around, though - it has potential
for other notifications (especially inter-user notifications), even if
setting up PMs would be pretty pointless.