Added structure for persistent messages
Basically, anything sent via the mailbox notifications API will now
spawn a notification that only goes away for good once it's clicked on.

I'll probably add more fields to the notification model at some point,
to make this work a bit better - being able to pass the notification
type and perhaps a boolean to indicate whether it needs to be
acknowledged would be useful.