Change deletion strategy, added publication option
Items now can be created without being published. Unpublished items are
only viewable by the author, and may be accessed from their profile page
(I may change this - letting editors view unpublished items is probably
a good idea - but it would require a bit more work in the controllers).

Also, nothing is truly deleted: comments and items are marked as deleted
in the database, and are then hidden from everyone. I'm thinking about
adding a way to see them (maybe an option for accounts with sufficient
permissions?) but for now they're just gone.