Adding group system
Added most of a group system; currently, it's barebones, and there's no
interface for group creation or management, but the bits related to
items work.

The goal for this is to allow authors to decide who has access to their
items; items can be only viewable by logged in users, or a subset of
logged in users. Probably not strictly necessary, and I doubt if it will
see much use, but potentially useful.